pyproject.toml
[project]
name = "basic-app"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "aiomysql>=0.3.2",
    "alembic>=1.18.4",
    "fastapi>=0.135.3",
    "greenlet>=3.3.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.12.1",
    "pymysql>=1.1.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-multipart>=0.0.22",
    "ruff>=0.15.9",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.43.0",
]


[tool.ruff]
line-length = 100
